+++ packages/components-react/projects/nextjs/tests/vrt/specs/radio-button-wrapper.vrt.ts
@@ -1,6 +1,6 @@
 import { defaultViewports, getVisualRegressionTester, vrtTest } from '@porsche-design-system/shared/testing';
 
-xit.each(defaultViewports)('should have no visual regression for viewport %s', async (viewport) => {
+it.each(defaultViewports)('should have no visual regression for viewport %s', async (viewport) => {
   expect(
     await vrtTest(getVisualRegressionTester(viewport), 'radio-button-wrapper', '/radio-button-wrapper', {
       javaScriptEnabled: false,